+++ src/modules/texts/TextEdit/components/DialogShare/services/types/fetchGetTextPermissions.types.ts
