+++ src/modules/texts/TextEdit/components/DialogShare/services/types/fetchPatchTextPermissions.types.ts
