+++ src/modules/texts/TextEdit/components/DialogShare/services/types/index.ts
